refactor(app): use proper loader component boundaries
